update core libraries to latest versions from fuglu

+++ src/postomaat/daemon.py
@@ -1,4 +1,4 @@
-#   Copyright 2009-2016 Oli Schacher
+#   Copyright 2009-2017 Oli Schacher
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -27,12 +27,15 @@ class DaemonStuff(object):
     def __init__(self, pidfilename):
         self.pidfile = pidfilename
 
+
     def delpid(self):
         """Delete the pid file"""
         try:
-            os.remove(self.pidfile)
-        except:
+            if os.path.exists(self.pidfile):
+                os.remove(self.pidfile)
+        except Exception:
             pass
+
 
     def createDaemon(self):
         """Detach a process from the controlling terminal and run it in the
@@ -87,6 +90,7 @@ class DaemonStuff(object):
         os.close(pidfd)
         return 0
 
+
     def drop_privs(self, username='nobody', groupname='nobody', keep_supplemental_groups=True):
         """Drop privileges of the current process to specified unprivileged user and group. If keep_supplemental_groups is True,
         the process will also be associated with all groups the unprivileged user belongs to.
@@ -94,7 +98,7 @@ class DaemonStuff(object):
         try:
             running_uid = pwd.getpwnam(username).pw_uid
             running_gid = grp.getgrnam(groupname).gr_gid
-        except:
+        except Exception:
             raise Exception('Can not drop privileges, user %s or group %s does not exist' % (
                 username, groupname))
         new_umask = 0o077
@@ -104,6 +108,7 @@ class DaemonStuff(object):
         if keep_supplemental_groups:
             os.setgroups(self._get_group_ids(username))
         os.setuid(running_uid)
+
 
     def _get_group_ids(self, username):
         """Return a list of group ids the user belongs to"""
